move Query\Builder to Post namespace as QueryBuilder
